k8s: Add CRD Identities as a identity allocator backend
The ciliumidentity CRD can act as the backing store for identity
allocation in cilium. Implementation details are hidden behind the
pkg/allocator.Backend interface. Most operations are supported but,
specifically, locking is not.

The name of the CRD is the numeric identity and the labels on the CRD
object are the security labels used by cilium. Each node using a
ciliumidentity updates the status field with it's name and a timestamp.
cilium-operator can then use the list of nodes in status to reference
count users of this identity, and to expire stale usage.

Both cilium and cilium-operator need full access to create, update and
delete these entities. cilium-operator does the garbage collection on
these objects and can also syncronize them to a KVStore.

Signed-off-by: Ray Bejjani <ray@covalent.io>