operator: support subnets filters
The cilium-operator start/resync time and memory consumption are excessive in our largest AWS accounts, due to the GetInstances regularly fetching each VPC's interfaces.
We run our clusters in dedicated subnets, so the set of relevant interfaces can be easily reduced.

In one of our AWS accounts, fetching all subnets and interfaces requires about 2GB of memory, and takes about 60s.
Filtering on cluster's subnets, we're down to 20MB and 2s for a 130 nodes cluster (or 200MB / 8s for 2300 nodes).