routing: Consistent conditions for ip rule creation/deletion
[ upstream commit 298ece34407caee7eeffacdac35a85af094ecb4b ]

In addition to the inconsistency fixed in the previous commit, there is
another we need to fix. On the deletion of ip rules with the form "to
x.x.x.x/x", we check that we are running in ENI mode. We don't perform
the same check on the creation of those rules. Thus, on AKS and
AlibabaCloud, we could theoretically create rules of one form and
attempt to delete the other form.

In practice, this isn't causing a bug because:
- The CIDR used in the "to x.x.x.x/x" form appears to always be
  0.0.0.0/0 and is therefore equivalent to not having a to CIDR
  specified.
- We don't support equivalent flags to --aws-release-excess-ips on AKS
  and AlibabaCloud and, as detailed in the previous commit, that flag is
  necessary to trigger the bug.

Thus, this commit is fixing an inconsistency in the code that doesn't
have any consequence today.

Signed-off-by: Paul Chaignon <paul@cilium.io>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>